Copy compose-material3 from 278ea749891 (without build scripts and stubs)

+++ compose/material3/material3/integration-tests/material3-catalog/src/main/java/androidx/compose/material3/catalog/library/model/Themes.kt
@@ -36,7 +36,7 @@ data class Theme(
         colorMode = ColorMode.values()[map.getValue(ColorModeKey).toInt()],
         expressiveThemeMode =
             ExpressiveThemeMode.values()[map.getValue(ExpressiveThemeModeKey).toInt()],
-        fontScale = map.getValue(FontScaleKey).toFloat(),
+        fontScale = map.getValue(FontScaleKey),
         fontScaleMode = FontScaleMode.values()[map.getValue(FontScaleModeKey).toInt()],
         textDirection = TextDirection.values()[map.getValue(TextDirectionKey).toInt()],
         showOnlyExpressiveComponents = map.getValue(ShowOnlyExpressiveComponents).toInt() != 0,
